+++ example/lib/examples/examples.dart
@@ -3,6 +3,7 @@ export 'databinding.dart';
 export 'databinding_artboards.dart';
 export 'databinding_images.dart';
 export 'databinding_lists.dart';
+export 'demo_scripting.dart';
 export 'events.dart';
 export 'hit_test_behaviour.dart';
 export 'inputs.dart';